added fix for worker table bugs
