+++ commcare_connect/opportunity/tables.py
@@ -1,5 +1,4 @@
 import itertools
-from datetime import timedelta
 
 import django_tables2 as tables
 from crispy_forms.helper import FormHelper
@@ -26,12 +25,8 @@ from commcare_connect.opportunity.models import (
     VisitValidationStatus,
 )
 from commcare_connect.users.models import User
-
-
-class OrgContextTable(tables.Table):
-    def __init__(self, *args, **kwargs):
-        self.org_slug = kwargs.pop("org_slug", None)
-        super().__init__(*args, **kwargs)
+from commcare_connect.utils.tables import OrgContextTable, IndexColumn, ClickableRowsTable, STOP_CLICK_PROPAGATION, \
+    DurationColumn
 
 
 class LearnStatusTable(OrgContextTable):
@@ -574,28 +569,7 @@ def header_with_tooltip(label, tooltip_text):
     )
 
 
-class IndexColumn(tables.Column):
-    def __init__(self, *args, **kwargs):
-        kwargs.setdefault("verbose_name", "#")
-        kwargs.setdefault("orderable", False)
-        kwargs.setdefault("empty_values", ())
-        super().__init__(*args, **kwargs)
-
-    def render(self, value, record, bound_column, bound_row, **kwargs):
-        table = bound_row._table
-        page = getattr(table, "page", None)
-        if page:
-            start_index = (page.number - 1) * page.paginator.per_page + 1
-        else:
-            start_index = 1
-        if not hasattr(table, "_row_counter") or getattr(table, "_row_counter_start", None) != start_index:
-            table._row_counter = itertools.count(start=start_index)
-            table._row_counter_start = start_index
-        value = next(table._row_counter)
-        return value
-
-
-class BaseOpportunityList(OrgContextTable):
+class BaseOpportunityList(ClickableRowsTable):
     stats_style = "underline underline-offset-2 justify-center"
 
     def __init__(self, *args, **kwargs):
@@ -657,6 +631,13 @@ class BaseOpportunityList(OrgContextTable):
             "end_date",
         )
 
+    def row_click_url(self, record):
+        return reverse("opportunity:detail", args=(self.org_slug, record.id))
+
+    def excluded_columns(self):
+        return ["actions"]
+
+
     def render_status(self, value):
         if value == 0:
             badge_class = "badge badge-sm bg-green-600/20 text-green-600"
@@ -703,7 +684,7 @@ class OpportunityTable(BaseOpportunityList):
     inactive_workers = tables.Column()
     pending_approvals = tables.Column()
     payments_due = tables.Column()
-    actions = tables.Column(empty_values=(), orderable=False, verbose_name="")
+    actions = tables.Column(empty_values=(), orderable=False, verbose_name="", attrs=STOP_CLICK_PROPAGATION)
 
     class Meta(BaseOpportunityList.Meta):
         sequence = BaseOpportunityList.Meta.sequence + (
@@ -771,7 +752,7 @@ class ProgramManagerOpportunityTable(BaseOpportunityList):
         verbose_name="Verified Deliveries"
     )
     worker_earnings = tables.Column(verbose_name="Worker Earnings", accessor="total_accrued")
-    actions = tables.Column(empty_values=(), orderable=False, verbose_name="")
+    actions = tables.Column(empty_values=(), orderable=False, verbose_name="", attrs=STOP_CLICK_PROPAGATION)
 
     class Meta(BaseOpportunityList.Meta):
         sequence = BaseOpportunityList.Meta.sequence + (
@@ -993,29 +974,6 @@ class SuspendedIndicatorColumn(tables.Column):
         return format_html('<div class="w-10"><div class="w-4 h-2 rounded {}"></div></div>', color_class)
 
 
-def get_duration_min(total_seconds):
-    total_seconds = int(total_seconds)
-    minutes = (total_seconds // 60) % 60
-    hours = (total_seconds // 3600) % 24
-    days = total_seconds // 86400
-
-    parts = []
-    if days:
-        parts.append(f"{days} day{'s' if days != 1 else ''}")
-    elif hours:
-        parts.append(f"{hours} hr")
-    elif minutes or not parts:
-        parts.append(f"{minutes} min")
-
-    return " ".join(parts)
-
-
-class DurationColumn(tables.Column):
-    def render(self, value):
-        total_seconds = int(value.total_seconds() if isinstance(value, timedelta) else 0)
-        return get_duration_min(total_seconds)
-
-
 class WorkerStatusTable(tables.Table):
     index = IndexColumn()
     user = UserInfoColumn()
@@ -1072,7 +1030,7 @@ class WorkerPaymentsTable(tables.Table):
         )
 
 
-class WorkerLearnTable(OrgContextTable):
+class WorkerLearnTable(ClickableRowsTable):
     index = IndexColumn()
     user = UserInfoColumn()
     suspended = SuspendedIndicatorColumn()
@@ -1121,6 +1079,9 @@ class WorkerLearnTable(OrgContextTable):
             "action",
         )
 
+    def row_click_url(self, record):
+        return reverse("opportunity:worker_learn_progress", args=(self.org_slug, self.opp_id, record.id))
+
     def render_assessment(self, value, record):
         if not record.date_learn_started:
             return "--"
@@ -1136,7 +1097,7 @@ class WorkerLearnTable(OrgContextTable):
         )
 
 
-class WorkerDeliveryTable(OrgContextTable):
+class WorkerDeliveryTable(ClickableRowsTable):
     use_view_url = True
 
     id = tables.Column(visible=False)
@@ -1175,11 +1136,15 @@ class WorkerDeliveryTable(OrgContextTable):
             "action",
         )
 
+
     def __init__(self, *args, **kwargs):
         self.opp_id = kwargs.pop("opp_id")
         self.use_view_url = True
         super().__init__(*args, **kwargs)
         self._seen_users = set()
+
+    def row_click_url(self, record):
+        return reverse("opportunity:user_visits_list", args=(self.org_slug, self.opp_id, record.id))
 
     def render_action(self, record):
         url = reverse("opportunity:user_visits_list", args=(self.org_slug, self.opp_id, record.id))
