+++ commcare_connect/templates/tailwind/base_table.html
@@ -31,7 +31,7 @@
                 {% block table.tbody.row %}
                   <tr class="group" {{ row.attrs.as_html }}>
                     {% for column, cell in row.items %}
-                      <td>
+                      <td {{ column.attrs.td.as_html }}>
                         {% if column.localize == None %}
                           {{ cell }}
                         {% else %}
